Rename logStrLength=>length; empty and null introduced
Making LogStr interface more similar to APIs of other string-like types.
Since length can be obtained in O(1), it is possible to define null with
the same complexity. This predicate can be used to create more precise
log messages. Another new function is "empty :: LogStr", which is now
used to define mempty for LogStr. It's basic advantage is that it is
monomorphic where mempty is not. Therefore in the situations when usage
of mempty would require type annotation, empty can be used instead.